+++ core/project.service.js
@@ -54,6 +54,25 @@ async function create_project(new_project, uuid, action_owner = 'auto') {
     catch(error) {
         logger.error(error);
     }
+    
+    const owner = await dbsrv.mongo_users().findOne({ uid: new_project.owner });
+    const msg_destinations =  [owner.email];
+    if (owner.send_copy_to_support) {
+        msg_destinations.push(CONFIG.general.support);
+    }
+    try {
+        await maisrv.send_notif_mail( {
+            'name': 'ask_project',
+            'destinations': msg_destinations,
+            'subject': 'Project ' + new_project.id + ' created'
+        }, {
+            '#NAME#': new_project.id,
+            '#PATH#': new_project.path,
+            '#GROUP#':new_project.group
+        });
+    } catch(error) {
+        logger.error(error);
+    }
     return new_project;
 }